# stage1_gateway_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ForgeAI - Api Gateway",
    "description": "API Gateway single public entry point providing authentication, resilience, and metrics.\n\n*Part of the ForgeAI Local AI Software Engineering Platform.*",
    "version": "0.1.0",
    "x-service-name": "api-gateway"
  },
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List Available LLM Models",
        "description": "Returns available Ollama coding models installed on the system.",
        "operationId": "list_models_api_v1_models_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Models list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_LLMModelInfo__"
                }
              }
            }
          },
          "503": {
            "description": "LLM Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/models/benchmark": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get Local Model Benchmark Report",
        "description": "Returns latency, throughput, RAM, VRAM, and JSON reliability benchmarks for evaluated local coding models.",
        "operationId": "get_model_benchmark_api_v1_models_benchmark_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Benchmark report retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ModelBenchmarkReport_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchmarkMetric": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model name"
          },
          "ram_mb": {
            "type": "number",
            "title": "Ram Mb",
            "description": "RAM consumption in MB"
          },
          "vram_mb": {
            "type": "number",
            "title": "Vram Mb",
            "description": "VRAM consumption in MB"
          },
          "ttft_ms": {
            "type": "number",
            "title": "Ttft Ms",
            "description": "Time to first token in milliseconds"
          },
          "tokens_per_sec": {
            "type": "number",
            "title": "Tokens Per Sec",
            "description": "Generation throughput tokens/sec"
          },
          "tool_reliability_score": {
            "type": "number",
            "title": "Tool Reliability Score",
            "description": "Tool calling accuracy percentage"
          },
          "json_reliability_score": {
            "type": "number",
            "title": "Json Reliability Score",
            "description": "Structured JSON compliance percentage"
          }
        },
        "type": "object",
        "required": [
          "model",
          "ram_mb",
          "vram_mb",
          "ttft_ms",
          "tokens_per_sec",
          "tool_reliability_score",
          "json_reliability_score"
        ],
        "title": "BenchmarkMetric",
        "description": "Benchmark performance entry for model comparison."
      },
...
      "ErrorDetails": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional context or validation errors"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorDetails",
        "description": "Detailed error info structure."
      },
      "ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "False for failed requests",
            "default": false
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code",
            "default": "INTERNAL_ERROR"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error description"
          },
          "service": {
            "type": "string",
            "title": "Service",
            "description": "Originating service name",
            "default": "api-gateway"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetails",
            "description": "Structured error details"
          }
        },
        "type": "object",
        "required": [
          "message",
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Unified error response wrapper conforming to production contract."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMModelInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Model identifier name",
            "example": "qwen2.5-coder:7b-instruct-q4_0"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human readable model name"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes",
            "description": "Model file size in bytes"
          },
          "parameter_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameter Size",
            "description": "Parameter count e.g. 7B, 3B"
          },
          "quantization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantization",
            "description": "Quantization e.g. Q4_0, Q8_0"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Currently selected model flag",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "LLMModelInfo",
        "description": "Model information schema."
      },
      "ModelBenchmarkReport": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Benchmark timestamp"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BenchmarkMetric"
            },
            "type": "array",
            "title": "Results"
          },
          "recommended_quality_model": {
            "type": "string",
            "title": "Recommended Quality Model",
            "description": "Recommended quality coding model"
          },
          "recommended_fast_model": {
            "type": "string",
            "title": "Recommended Fast Model",
            "description": "Recommended fast classification model"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "recommended_quality_model",
          "recommended_fast_model"
        ],
        "title": "ModelBenchmarkReport",
        "description": "Local benchmark execution report."
      },
...
      "SuccessResponse_ModelBenchmarkReport_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelBenchmarkReport"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[ModelBenchmarkReport]"
      },
...
      "SuccessResponse_list_LLMModelInfo__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "True for successful requests",
            "default": true
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID for request tracing",
            "default": "-"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "Distributed trace ID",
            "default": "-"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 UTC timestamp"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LLMModelInfo"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Payload data"
          }
        },
        "type": "object",
        "title": "SuccessResponse[list[LLMModelInfo]]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}